Extract WatsonxTextChatBase to share the text chat logic
